backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.25
psycopg[binary,pool]==3.2.1
pydantic==2.9.2
pydantic-settings==2.5.2
langchain==0.2.17
langchain-community==0.2.19
chromadb==0.5.5
ollama==0.2.1
httpx==0.27.2
alembic==1.13.3
python-dotenv==1.0.1
pytest==8.3.3
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
pdfrw==0.4
